Fix JBMC crash on windows in `NondetEnumOpaqueReturn` test
String literals with multibyte utf-8 characters in java cause `ch` to be
negative. Passing a negative value into `std::isalnum` results in
undefined behaviour. On windows the observed behaviour is a popup error
message about a failed assertion in the run time library. This stalls
the CI process as none of the buttons on the popup message ever get
pressed. Escaping these characters is sufficient to fix the stalling
test.