Remove line breaks from the body of the text on the release page.
The yaml file regular-release.yaml generating the release page include
the body of text for the release page.  The definition of the body
include line breaks to make the text fit within 80 characters in the
yaml file, but the line breaks in the yaml file are treated as line
breaks in the resulting markdown file, and the text does not flow well
in a web browser.

This page just removes the line breaks within a paragraph of text.