Add test directory for future sequentialization work
This directory of regression tests is for examples which we intend to
use for testing the future work on the analysis of concurrent programs
based on a sequentialization approach.

Tests which require POSIX specific headers will be disabled for the
moment because these are not natively available on Windows. A windows
POSIX compatability layer library could potentially be used to enable
them to be run on Windows. However the work to do so has not been done.