Define new `numberingt`
This commit defines a new `numberingt`, which is equivalent to the
existing `hash_numbering` type. This new name and its template
parameters match the cbmc coding standards by ending the type names with
a letter 't'. It also supports using `std::hash` by default, which will
saves passing a hash function in the case where there is a standard hash
function defined for a given type of `keyt`.

The intention is to add this new type in this commit, swap over
the existing usages of numbering in the subsequent commits and then
finally remove the old variations of numbering. This will improve
performance in one instance and will improve maintainability in others.
Maintainability will be improved with a single version, because there
won't be multiple versions to maintain and it will take away the
possibility of accidentally choosing a variation which uses `irept` as
the key in the backing `std::map`.