Use CCache to speed up build times in GitHub actions
We used to do this on Travis and CodeBuild, let's also port this to
GitHub actions. For any CMake builds, ccache will be used automatically
if available (as of 4b666920de).

For Visual Studio, we'd need https://github.com/ccache/ccache/pull/506
to be merged to use ccache.

On Ubuntu runners and OSX runners, the build time of a rebuild without
source code changes is down to under one minute (from 20 minutes).