String abstraction: add parameters to all function symbols
Do not restrict the updates to the functions in goto_functions, and
create parameters also for unnamed function parameters. While at it,
also cleanup the arguments vs. parameters terminology. Finally, string
abstraction (as it is currently implemented) really only works for a
complete goto model, the application to individual goto programs won't
really work. Thus deprecate the operator() methods in favour of a new
apply() method.