goto-symex: avoid redundant operations in pop_frame
There is no need to reconstruct the L1 object identifier when it is
already available as the key of a data structure being iterated over.
Also, perform the linear iteration over the smaller of the two
containers that we are computing the intersection of.

On a set of CSmith-generated tests (which 1) hardly require any use of a
SAT solver, and thus runtime is dominated by symex, and 2) contain a lot
of function calls), `perf` reported that 34% of CPU cycles were spent on
ssa_exprt::get_l1_object_identifier and its children. As stated above,
the call to this function in pop_frame is completely unnecessary. With
the changes applied, ssa_exprt::get_l1_object_identifier is below the 1%
mark.

For a different set of benchmarks: on
DeviceDriversLinux64Large-ReachSafety benchmarks, 9.59% of time were
spent on get_l1_object_identifier.