Simplification of pointer comparison: do not assume numeric types
As we cannot (or chose not to) catch all type inconsistencies in the
language front-end, we may pass expressions to simplification that do
not adhere to typing expectations. The included example is such a case,
where the missing declaration of `memcpy` makes the struct-to-pointer
cast get all the way to the middle end.

Test is based on a sample generated by C-Reduce when starting from an
SV-COMP task.