Remove unused simplify_exprt::objects_equal{,_address_of}
These methods are not documented are not called from anywhere ever since
ID_same_object was removed. As such, remove the declaration of
simplify_same_object (and that of simplify_dynamic_size) as well for
their implementations had been removed with the removal of
ID_same_object and ID_dynamic_size in b493cf2768.

Object comparisons are nowadays performed via an equality over
pointer_object expressions, which we have simplification rules for.