Eliminate array_abstract_objectt as a base class
Added aggregate_tag empty base class to abstract_aggregate_object so we
can verify for the absence of an aggregate class in certain
preconditions in value_set_abstract_objectt::get_type.

There's a stub value_set_array_abstract_object. I've dropped in
two_value_array_... as its base for now, but that will change when we
eventually fill that out.