Fix unpack_array_vector to produce an array of bytes
We wrongly created a return type (though not the underlying expression,
which was correct!) as an array of elements of the original subtype,
even though the sole purpose of unpack_array_vector is to turn the
original subtype into a sequence of bytes.

We seemingly neither had a test covering this code nor did we have
additional result verification in place. Both are now rectified: a new
regression test is added, and an invariant is checked right after
returning from unpack_rec (which may invoke unpack_array_vector).

Fixes: #5718