Isolate numeric version for image tag.
Previously, we were using the full release name
for generating a dockerhub image name, which if
it contained any characters like `-` would be creating
an issue.

This is now performing some preprocessing to isolate
the numeric part of the cbmc version (a string of the
form `cbmc-5.20.2`) so that it should be able to be
used for a tag name without any issues.