Replace boolbv_mapt::map_bitt by bvt
The Boolean flag is_set does not need to be stored for each individual
bit as all entries will either be false (initial state) or be true. It
is not possible for only some of the bits to be set.

On the benchmark that triggered me looking into this, memory consumption
caused by boolbv_mapt::get_map_entry was 4.50 GB, and is now 2.25 GB
(which is reasonable, because the bool-typed struct member likely
resulted in padding being generated that would effectively double the
memory consumption).

In addition to memory savings, some of the code can now be simplified as
the entire vector can be copied at once.