Ensure returned value_set objects have correct context
I'm not super thrilled about how I've done this, because ideally all
object creation should go through the object factory. However, in this
particular case a large value set could be converted to an interval. If
we hand off to the factory it'll just create another value-set, so we
need to explicitly create the interval. However, when we do that we
don't have the context wrapper (write location or data dependency).

Consequently, I've added variable_sensitivity_object_factoryt::wrap_with_context
and exposed it through abstract_environment so we can add the
appropriate context wrapper should we need it.