Added value_ranget value_range() to abstract_value_objectt
Iterates over the values represented by the abstract object. For
constants and intervals, this will be the object itself. For value sets,
it will be the values in the set.

Rework value_set expression evaluation in terms of value_ranget. This
enables evaluation with mixed value_sets and constants.

Made value_set_abstract_object implementation functions static.

There's a bodge in value_set_abstract_objectt::unwrap_operands to
workaround what would otherwise be an invariant violation. As we
implement cross-representation evaluation, we'll hit these exceptions
but once the work is complete they should all be eliminated.