Permit keeping trivially-true assertions
In some settings it is desirable to obtain a report of all properties
being checked, including those that are trivially true. There used to be
an option "all-claims", then renamed to "retain-trivial", but the latter
was never made available in command-line front-ends.

The option is now renamed to retain-trivial-checks, and made available
in all front-ends.