Explicitly create values with the required representation
When we create values through the factory, the representation used is
determined by the factory configuration. Up until now, that
configuration has always lined up with the representation - eg if we were
evaluating using intervals, then the factory is creating intervals.

This is no longer the case. As we move forward with cross-representation
evaluation, it's entirely possible that the factory configuration does
_not_ match the representation we have in hand during an evaluation.
Consequently, if we need a new constant value we should directly create
it.

As we leave the expression_transform, we use the factory to wrap any
required context around the returned object.