value_set index_range should flatten contained values into one range
A constant of, say, 5 has a index_range of 5. A value_set of { 5, 6,
9 } therefore has an index_range of 5,6,9.

An interval of [1, 3] has an index_range of 1,2,3. A value_set
containing that interval should have the same index_range.

A value_set of multiple constants and intervals should have an
index_range which is the union of its contents index_range.