Fix off-by-one error in unbounded byte update
The expression constructing individual elements within an array
comprehension did subtract the separately built first element, but
failed to consider that the index (and not number of elements) was being
used. Add 1 to the index to make it a "number of elements."

While at it, also fix typos in comments around this line of code.

Fixes: #5885