Use ccache in codecov-coverage-report GitHub action
The GitHub action takes by far the longest, and while the execution time
likely is dominated by test execution we should not unnecessarily waste
time for compiling.

Building previously took more than 20 minutes, and now takes around 3
minutes instead. The ccache size has to be set to at least 2 GB to avoid
cache entries being removed during build, but the packed archive of
cache results is below 500 MB, which should be safe as far as our use of
the 5 GB of total cache storage is concerned.