apply_condition: notequal need not have constant operand
Just as is done for equality, the not-equal case needs to test that the
LHS is a symbol and the RHS is a constant before attempting to constant
propagate.

The problem became apparent while trying to enable abort-after-failing-C
assert, but only on Visual Studio builds (owing to the code that the
assert macro expands to in Visual Studio).