Fix initialization of multiple sub-members within a union
The attached test shows that we could reach recursive initialisation of
unions with a byte_update instead of a union expression, which was added
in 1111e3110e. That previous fix did not consider the case of multiple
sub-members being initialised.

The test was constructed using creduce, starting from code included in
the Linux kernel.