introduce field_address_exprt and element_address_exprt
This commit introduces three new expressions, which fix the largest
asymmetry in the expression system: expressions that denote pointers to
objects.

The first expression is the base constructor: it returns the address of an
object with a given identifier.  It is semantically equivalent to the
existing address_of expression with a symbol subexpression.

The next two expressions both take a base address, and then add to it the
offset (in char units) of either a named struct field or an array element
(given by its index).  They are semantically equivalent to but much more
concise and readable than the equivalent combination of typecasts and
pointer arithmetic.

They key benefit is that algorithms that iterate over expressions no longer
need to special case the sub-expressions of address_of expressions.