Manage lifetime of MiniSat solver pointer using `unique_ptr`
This has the benefit of allowing the use of a default destructor rather
than a hand written destructor. Note that `satcheck_minisat2_baset` was
already non-copyable because its `solver_hardness` field was
non-copyable.

The default implementation of the destructor is in the `.cpp` rather
than the header file in order to avoid recompiling it every time the
header file is included.