Add tests of new command line option
So that the new option has some basic tests and a directory to place
tests for follow-up functionality.

Note that all tests are currently passed the `--slice-formula` argument
so that unimplemented functionality from `__cprover_initialise` does not
get passed to the solver if the user code does not depend on it. This
will support incremental implementation of each slice of work in doing
the implementation. Once the implementation is sufficiently complete to
support all expressions generated from `__cprover_initialise`, the
`--slice-formula` argument can be removed again.