Change timestamp server for the windows package codesign process.
It appears that `tsa.starfieldtech.com` is down for a couple of weeks
now, so we are changing the timestamp server to something that works
for now.

This is a hotfix to allow us to make a new release that works with
windows packages. A more robust solution is going to be developed
soon after this gets merged and unblocks us on the windows packages.