Fix doxygen "too many nodes" filter
The "too many nodes" warnings were previously filtered out by the python
file. However the regex in the python before this commit does not match
the exact message from the current version of doxygen. This commit
removes the expected warnings from the list file and makes the python
work as it was previously intended, by fixing the regex to match either
the current or previous warning texts.