Introduce euclidean_mod expression
This introduces a variant of the modulo operator, to be used on integers,
which uses Boute's Euclidean definition.  The result is never negative, and
thus differs from the C99 "%" operator.  The semantics match what SMT-LIB2
does in the integer theory.

The operator % is not defined by SMT-LIB2, and is removed from the parser.