Include `TEMPLATE_TEST_CASE`s in test count
So that we can use catch's `TEMPLATE_TEST_CASE` without the test count
failing when using make to run the tests. When using
`TEMPLATE_TEST_CASE` the number of tests added to catch's count of test
depends on the number of type arguments supplied. Therefore this update
is needed to count the arguments  to get the correct count.