Use capture by value in term to string conversion
Some of the outputs can currently be safely captured by value. However
there is the potential for accidentally introducing dangling reference
type bugs from changes elsewhere in the term to string conversion code.
Therefore this change will reduce the likelyhood of bugs being
introduced in future and should have minimal performance impact due to
the copy-on-write properties of ireps.