Simplify regex used in `is_valid_smt_identifier`
There are 2 simplifications here.
1) A swap to a normal string literal, rather that a raw string literal,
as there is only a single character which needs escaping. 2) A change to
remove the `^` and `$` for matching the start and end of the string.
These were redundant, because `std::regex_match` already requires the
regex to match the full string rather than a partial sub-match.

This simplification is already tested by the "smt_identifier_termt
construction" test in the smt_terms unit tests.