Handle empty unions via dedicated expression
GCC and Clang support empty unions, which the C standard does not
permit. We previously handled this special case by generating nil. This,
however, is not supported by the bitvector back-end. Rather than coming
up with selective handling of the otherwise unexpected nil, a proper
expression is introduced (and handled).

Fixes: #6163