Add template for node construction with validation propagation
For each non-leaf node in the `smt_responset` tree, the same logic is
required where the child nodes are validated and we construct the
non-leaf node if there are no errors, or collect and propagate the
errors upwards if there are errors validating the child nodes.

The template in this commit abstracts this logic out, so that it doesn't
need to be duplicated for each type of node we construct.