fix writeset inclusion checks on DEAD instructions
When handling DEAD instructions in function/loop assigns clause
instrumentation, the instruction iterator was not being incremented
correctly.
This led to instrumentation outside of the function/loop scope, and
spurious write set inclusion violations.

Moreover, for loops, declarations of some temporaries (those involved in
the "initialization" of loop counter, for instance) are "outside" of the
loop identified by CBMC, so we no longer raise an exception on not
finding a corresponding DECL for each DEAD.
These variables are writable because they appear as assigns clause
targets, not because they were local DECLs.