Remove the --mm option from goto-diff
This was tested for but does not appear as an option in
goto_diff_parse_options.h as a valid option so this code can never be used.
Worse, goto-diff doesn't include any of the code that uses it, nor can I
think what the memory model might have to do with diffing goto-programs.